+++ xtal-c/Classes/ASTNode.h
@@ -84,8 +84,8 @@ class ASTNode
 	typedef union
 		{
 		int64_t		intValue;		// For A_INTLIT, Value of the integer
-		int64_t 	identifier;		// For A_IDENT, Symbol slot number
-		int64_t		size;			// For A_SCALE, how much to scale by
+		int 	 	identifier;		// For A_IDENT, Symbol slot number
+		int			size;			// For A_SCALE, how much to scale by
 		} Value;
 	
 	/*************************************************************************\
